+++ CMakeLists.txt
@@ -9,9 +9,16 @@ set(SDK_INCLUDE ${SDK_HOME}/include)
 message("sdk home: ${SDK_HOME}")
 
 include_directories(${SDK_INCLUDE})
+include_directories(include)
+add_subdirectory(src)
 link_directories(${SDK_HOME})
 
-add_executable(MultiVector main.cpp)
+set(multi_vector_srcs
+        src/MultiVectorEngine.cpp)
+
+
+add_executable(MultiVector main.cpp
+               ${multi_vector_srcs})
 
 target_link_libraries(MultiVector
         milvus_sdk
